Creo nuevo dto para traer user ,ediante username+implementacion metod…

+++ BackEnd-IT-FitLab/src/main/java/org/iesalixar/daw2/ImanolTrespaderne/IT/FitLab/entities/User.java
@@ -56,7 +56,7 @@ public class User {
 
     @NotNull(message = "La altura es obligatoria.")
     @Column(name = "height", precision = 5, scale = 2, nullable = false)
-    private BigDecimal height;
+    private Integer height;
 
     @NotNull(message = "El peso es obligatorio.")
     @Column(name = "weight", precision = 5, scale = 2, nullable = false)
